update cli to work with new approach to tenants
