+++ cli.js
@@ -1,9 +1,36 @@
 #!/usr/bin/env node
 
+require('dotenv').config();
 const path = require("path");
+const fs = require("fs");
 const Prefixy = require(path.resolve(__dirname, "prefixy"));
 const program = require("commander");
 const jwt = require("jsonwebtoken");
+const CONFIG_FILE = path.resolve(__dirname, "prefixy-config.json");
+
+const updateTenant = tenant => {
+  let opts = {};
+
+    try {
+      opts = JSON.parse(fs.readFileSync(CONFIG_FILE, "utf8"));
+    } catch(e) {}
+
+    opts = { ...opts, tenant};
+
+    fs.writeFileSync(CONFIG_FILE, JSON.stringify(opts), "utf8");
+};
+
+const loadTenant = () => {
+  let opts = {};
+
+  try {
+    opts = JSON.parse(fs.readFileSync(CONFIG_FILE, "utf8"));
+  } catch(e) {}
+
+  return opts.tenant;
+};
+
+const tenant = loadTenant();
 
 program
   .version('0.0.1')
@@ -18,8 +45,8 @@ program
   .command('token <token>')
   .action(token => {
     try {
-      const tenant = jwt.verify(token, Prefixy.secret).tenant;
-      Prefixy.updateTenant(tenant);
+      const tenant = jwt.verify(token, process.env.SECRET).tenant;
+      updateTenant(tenant);
     } catch(e) {
       console.log("Invalid token -- tenant not updated");
       console.log(e);
@@ -30,7 +57,7 @@ program
   .command('tenant <tenant>')
   .action(tenant => {
     try {
-      Prefixy.updateTenantConfig(tenant);
+      updateTenant(tenant);
     } catch(e) {
       console.log("Invalid tenant -- not updated");
       console.log(e);
@@ -44,6 +71,7 @@ program
 
     try {
       result = await Prefixy.invoke(() => Prefixy.importFile(path));
+      Prefixy.quitRedisClient();
     } catch(e) {
       console.log(e);
     }
@@ -55,7 +83,8 @@ program
   .command('insert <completion>')
   .action(async completion => {
     try {
-      await Prefixy.invoke(() => Prefixy.insertCompletions([completion]));
+      await Prefixy.invoke(() => Prefixy.insertCompletions([completion], tenant));
+      Prefixy.quitRedisClient();
     } catch(e) {
       console.log(e);
     }
@@ -65,7 +94,8 @@ program
   .command('increment <completion>')
   .action(async completion => {
     try {
-      await Prefixy.invoke(() => Prefixy.increment(completion));
+      await Prefixy.invoke(() => Prefixy.increment(completion, tenant));
+      Prefixy.quitRedisClient();
     } catch(e) {
       console.log(e);
     }
@@ -78,11 +108,12 @@ program
   .action(async (prefixQuery, command) => {
     const withScores = command.withScores;
     const limit = command.limit;
-    const args = [prefixQuery, { withScores, limit }];
+    const args = [prefixQuery, tenant, { withScores, limit }];
     let result;
 
     try {
       result = await Prefixy.invoke(() => Prefixy.search(...args));
+      Prefixy.quitRedisClient();
     } catch(e) {
       console.log(e);
     }
@@ -94,32 +125,11 @@ program
   .command('delete <completion>')
   .action(async completion => {
     try {
-      await Prefixy.invoke(() => Prefixy.deleteCompletions([completion]));
+      await Prefixy.invoke(() => Prefixy.deleteCompletions([completion], tenant));
+      Prefixy.quitRedisClient();
     } catch(e) {
       console.log(e);
     }
   });
-
-program
-  .command('persist <prefix>')
-  .action(async prefix => {
-    try {
-      await Prefixy.invoke(() => Prefixy.persistPrefix(prefix));
-    } catch(e) {
-      console.log(e);
-    }
-  });
-
-program
-  .command('load <prefix>')
-  .action(async prefix => {
-    try {
-      await Prefixy.invoke(() => Prefixy.loadPrefix(prefix));
-    } catch(e) {
-      console.log(e);
-    }
-  });
-
 
 program.parse(process.argv);
-Prefixy.quitRedisClient();
