+++ prefixy-config.json
@@ -1,1 +1,1 @@
-{"tenant":"5cl0GK"}
+{"tenant":"rSewct"}
