+++ prefixy.js
@@ -79,18 +79,6 @@ class Prefixy {
     return { ...this.defaultOpts(), ...opts };
   }
 
-  updateTenantConfig(tenant) {
-    let opts = {};
-
-    try {
-      opts = JSON.parse(fs.readFileSync(CONFIG_FILE, "utf8"));
-    } catch(e) {}
-
-    opts = { ...opts, tenant};
-
-    fs.writeFileSync(CONFIG_FILE, JSON.stringify(opts), "utf8");
-  }
-
   async invoke(cb) {
     return cb()
       .then((result) => {
